backend/requirements.txt
# Modern TTS/STT Backend Requirements
# FastAPI for Audio Processing with Stable Dependencies

# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing (Modern & Stable)
soundfile>=0.12.1
numpy>=1.24.3
librosa>=0.10.1
pydub>=0.25.1

# Modern TTS/STT Engines
gtts>=2.4.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
openai-whisper>=20231117

# Alternative TTS (More Reliable)
edge-tts>=6.1.9
elevenlabs>=0.2.26

# Audio Enhancement
noisereduce>=2.0.1
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
aiofiles>=23.2.1

# Development & Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2 